requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
et_xmlfile==2.0.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jobspy==0.31.0
lxml==6.0.2
markdownify==0.13.1
numpy==1.26.3
ollama==0.6.1
openpyxl==3.1.5
pandas==2.3.3
playwright==1.58.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jobspy==1.1.82
python-telegram-bot==22.6
pytz==2025.2
redis==7.1.1
regex==2024.11.6
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
tls-client==1.0.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
aiohttp>=3.9.0
duckduckgo-search>=7.0.0
